Have common interface for standard rejections
... and use it for routing in tests.
