+++ server/src/test/java/io/spine/server/procman/given/repo/TestProcessManager.java
@@ -192,7 +192,6 @@ public class TestProcessManager
 
     @React
     Nothing on(StandardRejections.EntityAlreadyDeleted rejection) {
-        keep(rejection);
         return nothing();
     }
